Add files via upload
As data, we used a bank customer information set, in which the default flag was marked. To increase the accuracy of the models, we generated new relative features. Moreover, we removed features for which the Gini index was less than 5% due to the fact that they have a low predictive level. Then, we took an enterprise approach and performed client-level modeling using XAI. 
As a result, we came to the conclusion that at the moment the most prospective approach is SHAP, since it exceeds the rest with a set of additional functions and modern design, so it will be easy for a person who is not familiar with ML to interpret the result easily. However, the LIME method also has quite a lot of potential. The result of this method is the most convenient to understand, but visually it looks old-fashioned.